extracted getCardsByFilter controller
